textureUpdates
Decided to use the cement texture for any of the cement roads so they all match, imported the ramp textures but still need to check if there's a variant that matches the variant I used, since var2 of the textures does not match var2 of the ramp. Need someone to import the enemy models.